Trait poisson::algorithm::Algorithm

pub trait Algorithm<F, V> where
    F: Float,
    V: Vector<F>, {
    fn next<R>(&mut self, _: &mut Builder<F, V>, _: &mut R) -> Option<V>
    where
        R: Rng;

    fn size_hint(&self, _: &Builder<F, V>) -> (usize, Option<usize>);

    fn restrict(&mut self, _: V);

    fn stays_legal(&self, _: &Builder<F, V>, _: V) -> bool;
}

Trait that describes poisson-disk distribution generating algorithm.

Required Methods

fn next<R>(&mut self, _: &mut Builder<F, V>, _: &mut R) -> Option<V> where
R: Rng,

Generates new sample advancing the algorithm.

fn size_hint(&self, _: &Builder<F, V>) -> (usize, Option<usize>)

Returns lower and upper bound of the amount of samples remaining for the algorithm to generate.

fn restrict(&mut self, _: V)

Restricts the algorithm with an arbitary sample.

fn stays_legal(&self, _: &Builder<F, V>, _: V) -> bool

Checks if a sample is valid for the poisson-disk distribution generated thus far by the algorithm.