Struct poisson::Builder

pub struct Builder<F, V> where
    F: Float,
    V: Vector<F>,  { /* fields omitted */ }

Builder for the generator.

Methods

impl<V, F> Builder<F, V> where
F: Float,
V: Vector<F>,

fn with_radius(radius: F, poisson_type: Type) -> Self

New Builder with type of distribution and radius specified. The radius should be ]0, √2 / 2]

fn with_relative_radius(relative: F, poisson_type: Type) -> Self

New Builder with type of distribution and relative radius specified. The relative radius should be ]0, 1]

fn with_samples(samples: usize, relative: F, poisson_type: Type) -> Self

New Builder with type of distribution, approximate amount of samples and relative radius specified. The amount of samples should be larger than 0. The relative radius should be [0, 1]. For non-perioditic this is supported only for 2, 3 and 4 dimensional generation. For perioditic this is supported up to 8 dimensions.

fn radius(&self) -> F

Returns the radius of the generator.

fn poisson_type(&self) -> Type

Returns the type of the generator.

fn build<R, A>(self, rng: R, _algo: A) -> Generator<F, V, R, A> where
R: Rng,
A: Creator<F, V>,

Builds generator with random number generator and algorithm specified.